move non execution related profiling files to their own package
